Split db package into domain files and add WithTx helper
Split the monolithic db.go (624 lines) into focused files:
- schema.go: DDL and ALTER TABLE migrations
- tasks.go: task CRUD operations
- projects.go: project operations
- backends.go: backend operations
- config.go: config assembly and setters

Added WithTx transaction helper for atomic multi-step operations
with closure-based commit/rollback. Fixed gofmt violations in
db_test.go and migrate.go.
