+++ internal/db/db.go
@@ -2,17 +2,13 @@ package db
 
 import (
 	"database/sql"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/drn/argus/internal/config"
-	"github.com/drn/argus/internal/model"
 	_ "modernc.org/sqlite"
 )
 
@@ -88,545 +84,27 @@ func (d *DB) Close() error {
 	return d.conn.Close()
 }
 
-func (d *DB) createTables() error {
-	ddl := `
-		CREATE TABLE IF NOT EXISTS schema_version (
-			version INTEGER NOT NULL
-		);
-		CREATE TABLE IF NOT EXISTS tasks (
-			id         TEXT PRIMARY KEY,
-			name       TEXT NOT NULL,
-			status     TEXT NOT NULL DEFAULT 'pending',
-			project    TEXT NOT NULL DEFAULT '',
-			branch     TEXT NOT NULL DEFAULT '',
-			prompt     TEXT NOT NULL DEFAULT '',
-			backend    TEXT NOT NULL DEFAULT '',
-			worktree   TEXT NOT NULL DEFAULT '',
-			agent_pid  INTEGER NOT NULL DEFAULT 0,
-			session_id TEXT NOT NULL DEFAULT '',
-			pr_url     TEXT NOT NULL DEFAULT '',
-			created_at TEXT NOT NULL,
-			started_at TEXT NOT NULL DEFAULT '',
-			ended_at   TEXT NOT NULL DEFAULT ''
-		);
-		CREATE TABLE IF NOT EXISTS projects (
-			name                TEXT PRIMARY KEY,
-			path                TEXT NOT NULL,
-			branch              TEXT NOT NULL DEFAULT '',
-			backend             TEXT NOT NULL DEFAULT '',
-			sandbox_enabled     TEXT NOT NULL DEFAULT '',
-			sandbox_deny_read   TEXT NOT NULL DEFAULT '',
-			sandbox_extra_write TEXT NOT NULL DEFAULT ''
-		);
-		CREATE TABLE IF NOT EXISTS backends (
-			name           TEXT PRIMARY KEY,
-			command        TEXT NOT NULL,
-			prompt_flag    TEXT NOT NULL DEFAULT '',
-			resume_command TEXT NOT NULL DEFAULT ''
-		);
-		CREATE TABLE IF NOT EXISTS config (
-			key   TEXT PRIMARY KEY,
-			value TEXT NOT NULL
-		);
-	`
-	if _, err := d.conn.Exec(ddl); err != nil {
+// WithTx executes fn within a transaction, holding the DB mutex for the
+// duration. If fn returns an error, the transaction is rolled back; otherwise
+// it is committed.
+//
+// IMPORTANT: fn must operate on the provided *sql.Tx directly. It MUST NOT
+// call any *DB methods (Get, Update, Tasks, etc.) as d.mu is held for the
+// duration and Go's sync.Mutex is not reentrant — doing so will deadlock.
+func (d *DB) WithTx(fn func(tx *sql.Tx) error) error {
+	d.mu.Lock()
+	defer d.mu.Unlock()
+
+	tx, err := d.conn.Begin()
+	if err != nil {
+		return fmt.Errorf("begin tx: %w", err)
+	}
+	defer tx.Rollback() //nolint:errcheck
+
+	if err := fn(tx); err != nil {
 		return err
 	}
-
-	// Add per-project sandbox columns to existing databases (safe to call multiple times;
-	// errors for already-existing columns are silently ignored).
-	for _, def := range []string{
-		"sandbox_enabled     TEXT NOT NULL DEFAULT ''",
-		"sandbox_deny_read   TEXT NOT NULL DEFAULT ''",
-		"sandbox_extra_write TEXT NOT NULL DEFAULT ''",
-	} {
-		d.conn.Exec(`ALTER TABLE projects ADD COLUMN ` + def) //nolint:errcheck
-	}
-
-	// Add pr_url column to existing tasks tables.
-	d.conn.Exec(`ALTER TABLE tasks ADD COLUMN pr_url TEXT NOT NULL DEFAULT ''`) //nolint:errcheck
-
-	// Add archived column to existing tasks tables.
-	d.conn.Exec(`ALTER TABLE tasks ADD COLUMN archived INTEGER NOT NULL DEFAULT 0`) //nolint:errcheck
-
-	// Add todo_path column to existing tasks tables.
-	d.conn.Exec(`ALTER TABLE tasks ADD COLUMN todo_path TEXT NOT NULL DEFAULT ''`) //nolint:errcheck
-	// Index for TasksByTodoPath queries (called on every tick).
-	d.conn.Exec(`CREATE INDEX IF NOT EXISTS idx_tasks_todo_path ON tasks(todo_path)`) //nolint:errcheck
-
-	// Add resume_command column to existing backends tables.
-	d.conn.Exec(`ALTER TABLE backends ADD COLUMN resume_command TEXT NOT NULL DEFAULT ''`) //nolint:errcheck
-
-	// KB FTS5 full-text search table (virtual table — CREATE VIRTUAL TABLE).
-	// Note: FTS5 doesn't support UPDATE; use DELETE+INSERT in a transaction.
-	if _, err := d.conn.Exec(`
-		CREATE VIRTUAL TABLE IF NOT EXISTS kb_documents USING fts5(
-			path UNINDEXED,
-			title,
-			body,
-			tags,
-			tier UNINDEXED,
-			tokenize = 'porter unicode61'
-		)
-	`); err != nil {
-		return fmt.Errorf("creating kb_documents fts5 table: %w", err)
-	}
-
-	// KB metadata table for non-text fields not suitable for FTS5.
-	if _, err := d.conn.Exec(`
-		CREATE TABLE IF NOT EXISTS kb_metadata (
-			path        TEXT PRIMARY KEY,
-			modified_at INTEGER NOT NULL,
-			ingested_at INTEGER NOT NULL,
-			word_count  INTEGER NOT NULL DEFAULT 0,
-			tier        TEXT NOT NULL DEFAULT 'hot'
-		)
-	`); err != nil {
-		return fmt.Errorf("creating kb_metadata table: %w", err)
-	}
-
-	// KB pending tasks table for vault task imports awaiting approval.
-	if _, err := d.conn.Exec(`
-		CREATE TABLE IF NOT EXISTS kb_pending_tasks (
-			id          INTEGER PRIMARY KEY AUTOINCREMENT,
-			name        TEXT NOT NULL,
-			project     TEXT NOT NULL DEFAULT '',
-			source_file TEXT NOT NULL,
-			created_at  TEXT NOT NULL,
-			UNIQUE(source_file, name)
-		)
-	`); err != nil {
-		return fmt.Errorf("creating kb_pending_tasks table: %w", err)
-	}
-
-	return nil
-}
-
-// --- Tasks ---
-
-// taskColumns is the canonical column list for task queries.
-const taskColumns = `id, name, status, project, branch, prompt, backend, worktree, agent_pid, session_id, pr_url, todo_path, archived, created_at, started_at, ended_at`
-
-// scanner is implemented by both *sql.Row and *sql.Rows.
-type scanner interface {
-	Scan(dest ...any) error
-}
-
-// scanTask reads a task from a row using the canonical column order.
-func scanTask(row scanner) (*model.Task, error) {
-	t := &model.Task{}
-	var status, createdAt, startedAt, endedAt string
-	var archived int
-	if err := row.Scan(&t.ID, &t.Name, &status, &t.Project, &t.Branch, &t.Prompt, &t.Backend, &t.Worktree, &t.AgentPID, &t.SessionID, &t.PRURL, &t.TodoPath, &archived, &createdAt, &startedAt, &endedAt); err != nil {
-		return nil, err
-	}
-	t.Status, _ = model.ParseStatus(status)
-	t.Archived = archived != 0
-	t.CreatedAt = parseTime(createdAt)
-	t.StartedAt = parseTime(startedAt)
-	t.EndedAt = parseTime(endedAt)
-	return t, nil
-}
-
-func (d *DB) Tasks() []*model.Task {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	rows, err := d.conn.Query(`SELECT ` + taskColumns + ` FROM tasks ORDER BY created_at ASC`)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	var tasks []*model.Task
-	for rows.Next() {
-		if t, err := scanTask(rows); err == nil {
-			tasks = append(tasks, t)
-		}
-	}
-	return tasks
-}
-
-func (d *DB) Add(t *model.Task) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	if t.ID == "" {
-		t.ID = generateID()
-	}
-	if t.CreatedAt.IsZero() {
-		t.CreatedAt = time.Now()
-	}
-
-	archivedInt := 0
-	if t.Archived {
-		archivedInt = 1
-	}
-	_, err := d.conn.Exec(`INSERT INTO tasks (id, name, status, project, branch, prompt, backend, worktree, agent_pid, session_id, pr_url, todo_path, archived, created_at, started_at, ended_at) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-		t.ID, t.Name, t.Status.String(), t.Project, t.Branch, t.Prompt, t.Backend, t.Worktree, t.AgentPID, t.SessionID, t.PRURL, t.TodoPath, archivedInt,
-		formatTime(t.CreatedAt), formatTime(t.StartedAt), formatTime(t.EndedAt))
-	return err
-}
-
-func (d *DB) Update(t *model.Task) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	archivedInt := 0
-	if t.Archived {
-		archivedInt = 1
-	}
-	res, err := d.conn.Exec(`UPDATE tasks SET name=?, status=?, project=?, branch=?, prompt=?, backend=?, worktree=?, agent_pid=?, session_id=?, pr_url=?, todo_path=?, archived=?, created_at=?, started_at=?, ended_at=? WHERE id=?`,
-		t.Name, t.Status.String(), t.Project, t.Branch, t.Prompt, t.Backend, t.Worktree, t.AgentPID, t.SessionID, t.PRURL, t.TodoPath, archivedInt,
-		formatTime(t.CreatedAt), formatTime(t.StartedAt), formatTime(t.EndedAt), t.ID)
-	if err != nil {
-		return err
-	}
-	n, _ := res.RowsAffected()
-	if n == 0 {
-		return fmt.Errorf("task not found: %s", t.ID)
-	}
-	return nil
-}
-
-func (d *DB) Delete(id string) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	res, err := d.conn.Exec(`DELETE FROM tasks WHERE id=?`, id)
-	if err != nil {
-		return err
-	}
-	n, _ := res.RowsAffected()
-	if n == 0 {
-		return fmt.Errorf("task not found: %s", id)
-	}
-	return nil
-}
-
-func (d *DB) Get(id string) (*model.Task, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	row := d.conn.QueryRow(`SELECT `+taskColumns+` FROM tasks WHERE id=?`, id)
-	t, err := scanTask(row)
-	if errors.Is(err, sql.ErrNoRows) {
-		return nil, fmt.Errorf("task not found: %s", id)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return t, nil
-}
-
-// TasksByTodoPath returns a map from todo_path to the most recent task with that path.
-// Only tasks with a non-empty todo_path are included. Ordered by created_at ASC so
-// later tasks overwrite earlier ones (most recent wins).
-func (d *DB) TasksByTodoPath() map[string]*model.Task {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	rows, err := d.conn.Query(`SELECT ` + taskColumns + ` FROM tasks WHERE todo_path != '' ORDER BY created_at ASC`)
-	if err != nil {
-		return nil
-	}
-	defer rows.Close()
-
-	m := make(map[string]*model.Task)
-	for rows.Next() {
-		if t, err := scanTask(rows); err == nil {
-			m[t.TodoPath] = t
-		}
-	}
-	return m
-}
-
-func (d *DB) PruneCompleted() ([]*model.Task, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	// Fetch completed tasks first
-	rows, err := d.conn.Query(`SELECT ` + taskColumns + ` FROM tasks WHERE status='complete'`)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var pruned []*model.Task
-	for rows.Next() {
-		if t, err := scanTask(rows); err == nil {
-			pruned = append(pruned, t)
-		}
-	}
-
-	if len(pruned) == 0 {
-		return nil, nil
-	}
-
-	_, err = d.conn.Exec(`DELETE FROM tasks WHERE status='complete'`)
-	if err != nil {
-		return nil, err
-	}
-	return pruned, nil
-}
-
-// WorktreePaths returns the set of all non-empty worktree paths currently in the DB.
-// Returns an error if the query fails — callers should skip orphan sweep on error
-// to avoid treating all worktrees as orphans.
-func (d *DB) WorktreePaths() (map[string]bool, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	rows, err := d.conn.Query(`SELECT worktree FROM tasks WHERE worktree != ''`)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	paths := make(map[string]bool)
-	for rows.Next() {
-		var p string
-		if err := rows.Scan(&p); err == nil {
-			paths[p] = true
-		}
-	}
-	return paths, nil
-}
-
-// --- Projects ---
-
-func (d *DB) Projects() map[string]config.Project {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	rows, err := d.conn.Query(`SELECT name, path, branch, backend, sandbox_enabled, sandbox_deny_read, sandbox_extra_write FROM projects ORDER BY name`)
-	if err != nil {
-		return make(map[string]config.Project)
-	}
-	defer rows.Close()
-
-	projects := make(map[string]config.Project)
-	for rows.Next() {
-		var name string
-		var p config.Project
-		var sandboxEnabled, sandboxDenyRead, sandboxExtraWrite string
-		if err := rows.Scan(&name, &p.Path, &p.Branch, &p.Backend, &sandboxEnabled, &sandboxDenyRead, &sandboxExtraWrite); err != nil {
-			continue
-		}
-		switch sandboxEnabled {
-		case "true":
-			v := true
-			p.Sandbox.Enabled = &v
-		case "false":
-			v := false
-			p.Sandbox.Enabled = &v
-		}
-		if sandboxDenyRead != "" {
-			p.Sandbox.DenyRead = splitCSV(sandboxDenyRead)
-		}
-		if sandboxExtraWrite != "" {
-			p.Sandbox.ExtraWrite = splitCSV(sandboxExtraWrite)
-		}
-		projects[name] = p
-	}
-	return projects
-}
-
-func (d *DB) SetProject(name string, p config.Project) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	sandboxEnabled := ""
-	if p.Sandbox.Enabled != nil {
-		if *p.Sandbox.Enabled {
-			sandboxEnabled = "true"
-		} else {
-			sandboxEnabled = "false"
-		}
-	}
-	sandboxDenyRead := strings.Join(p.Sandbox.DenyRead, ",")
-	sandboxExtraWrite := strings.Join(p.Sandbox.ExtraWrite, ",")
-
-	_, err := d.conn.Exec(`INSERT OR REPLACE INTO projects (name, path, branch, backend, sandbox_enabled, sandbox_deny_read, sandbox_extra_write) VALUES (?, ?, ?, ?, ?, ?, ?)`,
-		name, p.Path, p.Branch, p.Backend, sandboxEnabled, sandboxDenyRead, sandboxExtraWrite)
-	return err
-}
-
-func (d *DB) DeleteProject(name string) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	_, err := d.conn.Exec(`DELETE FROM projects WHERE name=?`, name)
-	return err
-}
-
-// --- Backends ---
-
-func (d *DB) Backends() map[string]config.Backend {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	rows, err := d.conn.Query(`SELECT name, command, prompt_flag FROM backends ORDER BY name`)
-	if err != nil {
-		return make(map[string]config.Backend)
-	}
-	defer rows.Close()
-
-	backends := make(map[string]config.Backend)
-	for rows.Next() {
-		var name string
-		var b config.Backend
-		if err := rows.Scan(&name, &b.Command, &b.PromptFlag); err != nil {
-			continue
-		}
-		backends[name] = b
-	}
-	return backends
-}
-
-func (d *DB) SetBackend(name string, b config.Backend) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	_, err := d.conn.Exec(`INSERT OR REPLACE INTO backends (name, command, prompt_flag) VALUES (?, ?, ?)`,
-		name, b.Command, b.PromptFlag)
-	return err
-}
-
-// --- Config (assembles config.Config from DB) ---
-
-func (d *DB) Config() config.Config {
-	cfg := config.DefaultConfig()
-
-	// Load backends
-	cfg.Backends = d.Backends()
-
-	// Load projects
-	cfg.Projects = d.Projects()
-
-	// Load scalar config values — hold mutex through iteration
-	// to prevent concurrent writes while the rows cursor is open.
-	d.mu.Lock()
-	kv := make(map[string]string)
-	rows, err := d.conn.Query(`SELECT key, value FROM config`)
-	if err != nil {
-		d.mu.Unlock()
-		return cfg
-	}
-	for rows.Next() {
-		var k, v string
-		if err := rows.Scan(&k, &v); err != nil {
-			continue
-		}
-		kv[k] = v
-	}
-	rows.Close()
-	d.mu.Unlock()
-
-	// String config fields: map config key → pointer to struct field.
-	stringFields := []struct {
-		key  string
-		dest *string
-	}{
-		{"defaults.backend", &cfg.Defaults.Backend},
-		{"keybindings.new", &cfg.Keybindings.New},
-		{"keybindings.attach", &cfg.Keybindings.Attach},
-		{"keybindings.status", &cfg.Keybindings.Status},
-		{"keybindings.delete", &cfg.Keybindings.Delete},
-		{"keybindings.quit", &cfg.Keybindings.Quit},
-		{"keybindings.help", &cfg.Keybindings.Help},
-		{"keybindings.filter", &cfg.Keybindings.Filter},
-		{"keybindings.prompt", &cfg.Keybindings.Prompt},
-		{"keybindings.worktree", &cfg.Keybindings.Worktree},
-		{"ui.theme", &cfg.UI.Theme},
-	}
-	for _, f := range stringFields {
-		if v, ok := kv[f.key]; ok {
-			*f.dest = v
-		}
-	}
-
-	// Bool config fields
-	boolFields := []struct {
-		key  string
-		dest *bool
-	}{
-		{"ui.show_elapsed", &cfg.UI.ShowElapsed},
-		{"ui.show_icons", &cfg.UI.ShowIcons},
-	}
-	for _, f := range boolFields {
-		if v, ok := kv[f.key]; ok {
-			*f.dest = v == "true"
-		}
-	}
-
-	// Optional bool (pointer) config fields
-	if v, ok := kv["ui.cleanup_worktrees"]; ok {
-		val := v == "true"
-		cfg.UI.CleanupWorktrees = &val
-	}
-
-	// Sandbox config
-	if v, ok := kv["sandbox.enabled"]; ok {
-		cfg.Sandbox.Enabled = v == "true"
-	}
-	if v, ok := kv["sandbox.deny_read"]; ok && v != "" {
-		cfg.Sandbox.DenyRead = splitCSV(v)
-	}
-	if v, ok := kv["sandbox.extra_write"]; ok && v != "" {
-		cfg.Sandbox.ExtraWrite = splitCSV(v)
-	}
-
-	// KB config
-	if v, ok := kv["kb.enabled"]; ok {
-		cfg.KB.Enabled = v == "true"
-	}
-	if v, ok := kv["kb.http_port"]; ok {
-		if port, err := strconv.Atoi(v); err == nil && port > 0 {
-			cfg.KB.HTTPPort = port
-		}
-	}
-	if v, ok := kv["kb.metis_vault_path"]; ok {
-		cfg.KB.MetisVaultPath = v
-	}
-	if v, ok := kv["kb.argus_vault_path"]; ok {
-		cfg.KB.ArgusVaultPath = v
-	}
-	if v, ok := kv["kb.auto_create_tasks"]; ok {
-		cfg.KB.AutoCreateTasks = v == "true"
-	}
-
-	return cfg
-}
-
-// splitCSV splits a comma-separated string into trimmed non-empty parts.
-func splitCSV(s string) []string {
-	parts := strings.Split(s, ",")
-	var result []string
-	for _, p := range parts {
-		p = strings.TrimSpace(p)
-		if p != "" {
-			result = append(result, p)
-		}
-	}
-	return result
-}
-
-func (d *DB) SetConfigValue(key, value string) error {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	_, err := d.conn.Exec(`INSERT OR REPLACE INTO config (key, value) VALUES (?, ?)`, key, value)
-	return err
-}
-
-// SetSandboxEnabled toggles sandbox mode.
-func (d *DB) SetSandboxEnabled(enabled bool) error {
-	v := "false"
-	if enabled {
-		v = "true"
-	}
-	return d.SetConfigValue("sandbox.enabled", v)
+	return tx.Commit()
 }
 
 // --- Helpers ---
@@ -648,4 +126,17 @@ func parseTime(s string) time.Time {
 
 func generateID() string {
 	return fmt.Sprintf("%d", time.Now().UnixNano())
+}
+
+// splitCSV splits a comma-separated string into trimmed non-empty parts.
+func splitCSV(s string) []string {
+	parts := strings.Split(s, ",")
+	var result []string
+	for _, p := range parts {
+		p = strings.TrimSpace(p)
+		if p != "" {
+			result = append(result, p)
+		}
+	}
+	return result
 }
