+++ internal/db/migrate.go
@@ -67,23 +67,23 @@ func (d *DB) seedDefaults() error {
 	d.conn.QueryRow(`SELECT COUNT(*) FROM config`).Scan(&configCount)
 	if configCount == 0 {
 		defaults := map[string]string{
-			"defaults.backend":      cfg.Defaults.Backend,
-			"keybindings.new":       cfg.Keybindings.New,
-			"keybindings.attach":    cfg.Keybindings.Attach,
-			"keybindings.status":    cfg.Keybindings.Status,
-			"keybindings.delete":    cfg.Keybindings.Delete,
-			"keybindings.quit":      cfg.Keybindings.Quit,
-			"keybindings.help":      cfg.Keybindings.Help,
-			"keybindings.filter":    cfg.Keybindings.Filter,
-			"keybindings.prompt":    cfg.Keybindings.Prompt,
-			"keybindings.worktree":  cfg.Keybindings.Worktree,
-			"ui.theme":              cfg.UI.Theme,
-			"ui.show_elapsed":       fmt.Sprintf("%t", cfg.UI.ShowElapsed),
-			"ui.show_icons":         fmt.Sprintf("%t", cfg.UI.ShowIcons),
-			"kb.http_port":          fmt.Sprintf("%d", cfg.KB.HTTPPort),
-			"kb.metis_vault_path":   config.DefaultMetisVaultPath(),
-			"kb.argus_vault_path":   config.DefaultArgusVaultPath(),
-			"kb.auto_create_tasks":  fmt.Sprintf("%t", cfg.KB.AutoCreateTasks),
+			"defaults.backend":     cfg.Defaults.Backend,
+			"keybindings.new":      cfg.Keybindings.New,
+			"keybindings.attach":   cfg.Keybindings.Attach,
+			"keybindings.status":   cfg.Keybindings.Status,
+			"keybindings.delete":   cfg.Keybindings.Delete,
+			"keybindings.quit":     cfg.Keybindings.Quit,
+			"keybindings.help":     cfg.Keybindings.Help,
+			"keybindings.filter":   cfg.Keybindings.Filter,
+			"keybindings.prompt":   cfg.Keybindings.Prompt,
+			"keybindings.worktree": cfg.Keybindings.Worktree,
+			"ui.theme":             cfg.UI.Theme,
+			"ui.show_elapsed":      fmt.Sprintf("%t", cfg.UI.ShowElapsed),
+			"ui.show_icons":        fmt.Sprintf("%t", cfg.UI.ShowIcons),
+			"kb.http_port":         fmt.Sprintf("%d", cfg.KB.HTTPPort),
+			"kb.metis_vault_path":  config.DefaultMetisVaultPath(),
+			"kb.argus_vault_path":  config.DefaultArgusVaultPath(),
+			"kb.auto_create_tasks": fmt.Sprintf("%t", cfg.KB.AutoCreateTasks),
 		}
 		for k, v := range defaults {
 			if _, err := d.conn.Exec(`INSERT OR IGNORE INTO config (key, value) VALUES (?, ?)`, k, v); err != nil {
